+++ cisco_sdwan/tasks/implementation/_backup.py
@@ -1,19 +1,19 @@
 import argparse
-from typing import Union, Optional, List
-from pydantic import validator
+from typing import Union, Optional, List, Dict, Any
+from pydantic import validator, root_validator
+from uuid import uuid4
 from cisco_sdwan.__version__ import __doc__ as title
 from cisco_sdwan.base.rest_api import Rest, RestAPIException
 from cisco_sdwan.base.catalog import catalog_iter, CATALOG_TAG_ALL
-from cisco_sdwan.base.models_base import ServerInfo, SASTRE_ROOT_DIR, DATA_DIR
+from cisco_sdwan.base.models_base import ServerInfo
 from cisco_sdwan.base.models_vmanage import (DeviceConfig, DeviceConfigRFS, DeviceTemplate, DeviceTemplateAttached,
                                              DeviceTemplateValues, EdgeInventory, ControlInventory, EdgeCertificate,
                                              ConfigGroup, ConfigGroupValues, ConfigGroupAssociated, ConfigGroupRules)
 from cisco_sdwan.tasks.utils import TaskOptions, TagOptions, filename_type, regex_type, default_workdir
-from cisco_sdwan.tasks.common import regex_search, clean_dir, Task
-from cisco_sdwan.tasks.models import TaskArgs, validate_regex, validate_filename, validate_catalog_tag
-import pathlib
-from zipfile import *
-import os
+from cisco_sdwan.tasks.common import regex_search, clean_dir, Task, archive_create
+from cisco_sdwan.tasks.models import TaskArgs, validate_catalog_tag
+from cisco_sdwan.tasks.validators import validate_regex, validate_filename
+
 
 @TaskOptions.register('backup')
 class TaskBackup(Task):
@@ -23,6 +23,12 @@ class TaskBackup(Task):
         task_parser.prog = f'{task_parser.prog} backup'
         task_parser.formatter_class = argparse.RawDescriptionHelpFormatter
 
+        mutex_dest = task_parser.add_mutually_exclusive_group()
+        mutex_dest.add_argument('--archive', metavar='<filename>', type=filename_type,
+                                help='backup to zip archive')
+        mutex_dest.add_argument('--workdir', metavar='<directory>', type=filename_type,
+                                default=default_workdir(target_address),
+                                help='backup to directory (default: %(default)s)')
         task_parser.add_argument('--no-rollover', action='store_true',
                                  help='by default, if workdir already exists (before a new backup is saved) the old '
                                       'workdir is renamed using a rolling naming scheme. This option disables this '
@@ -30,15 +36,11 @@ class TaskBackup(Task):
         task_parser.add_argument('--save-running', action='store_true',
                                  help='include the running config from each node to the backup. This is useful for '
                                       'reference or documentation purposes. It is not needed by the restore task.')
-        mutex = task_parser.add_mutually_exclusive_group()
-        mutex.add_argument('--regex', metavar='<regex>', type=regex_type,
-                           help='regular expression matching item names to backup, within selected tags.')
-        mutex.add_argument('--not-regex', metavar='<regex>', type=regex_type,
-                           help='regular expression matching item names NOT to backup, within selected tags.')
-         mutex.add_argument('--workdir', metavar='<directory>', type=filename_type, default=default_workdir(target_address),
-                           help='backup destination (default: %(default)s)')
-        mutex.add_argument('--zip', metavar='<zip_file_path>.zip', type=zip, default=f'{default_workdir(target_address)}.zip',
-                           help='backup zipfile name (default: %(default_workdir(target_address)s).zip')
+        mutex_regex = task_parser.add_mutually_exclusive_group()
+        mutex_regex.add_argument('--regex', metavar='<regex>', type=regex_type,
+                                 help='regular expression matching item names to backup, within selected tags.')
+        mutex_regex.add_argument('--not-regex', metavar='<regex>', type=regex_type,
+                                 help='regular expression matching item names NOT to backup, within selected tags.')
         task_parser.add_argument('tags', metavar='<tag>', nargs='+', type=TagOptions.tag,
                                  help='one or more tags for selecting items to be backed up. Multiple tags should be '
                                       f'separated by space. Available tags: {TagOptions.options()}. Special tag '
@@ -47,7 +49,12 @@ class TaskBackup(Task):
         return task_parser.parse_args(task_args)
 
     def runner(self, parsed_args, api: Optional[Rest] = None) -> Union[None, list]:
-        self.log_info(f'Backup task: vManage URL: "{api.base_url}" -> Local workdir: "{parsed_args.workdir}"')
+        if parsed_args.archive:
+            self.log_info(f'Backup task: vManage URL: "{api.base_url}" -> Local archive file: "{parsed_args.archive}"')
+            parsed_args.workdir = str(uuid4())
+            self.log_debug(f'Temporary workdir: {parsed_args.workdir}')
+        else:
+            self.log_info(f'Backup task: vManage URL: "{api.base_url}" -> Local workdir: "{parsed_args.workdir}"')
 
         # Backup workdir must be empty for a new backup
         saved_workdir = clean_dir(parsed_args.workdir, max_saved=0 if parsed_args.no_rollover else 99)
@@ -123,20 +130,13 @@ class TaskBackup(Task):
                             continue
                         if sub_item.save(parsed_args.workdir, item_index.need_extended_name, item_name, item_id):
                             self.log_info(f'Done {info} {item_name} {sub_item_info}')
-                            
-        # If --zip option is passed by the user, create a .zip file for the backup
-        if parsed_args.zip or parsed_args.workdir:
-            self.file_zipper(parsed_args.workdir)
-            self.log_info('Backup zip file "%s.zip" created', parsed_args.workdir)
-        return
-      
-  @staticmethod
-    def file_zipper(workdir: str) -> None:
-        backup_regular_directory = pathlib.Path(DATA_DIR, f'{workdir}')
-        backup_zipfile_directory = pathlib.Path(SASTRE_ROOT_DIR, f'{workdir}.zip')
-        with ZipFile(backup_zipfile_directory, mode="w") as archive:
-            for file_path in backup_regular_directory.rglob("*"):
-                archive.write(file_path, arcname=file_path.relative_to(backup_regular_directory))
+
+        if parsed_args.archive:
+            archive_create(parsed_args.archive, parsed_args.workdir)
+            self.log_info(f'Created archive file "{parsed_args.archive}"')
+            clean_dir(parsed_args.workdir, max_saved=0)
+            self.log_debug('Temporary workdir deleted')
+
         return
 
     def save_running_configs(self, api: Optional[Rest], workdir: str) -> None:
@@ -164,15 +164,25 @@ class TaskBackup(Task):
 
 
 class BackupArgs(TaskArgs):
+    archive: Optional[str] = None
+    workdir: Optional[str] = None
     regex: Optional[str] = None
     not_regex: Optional[str] = None
     no_rollover: bool = False
     save_running: bool = False
-    zip: bool = False
-    workdir: str
     tags: List[str]
 
     # Validators
+    _validate_filename = validator('workdir', 'archive', allow_reuse=True)(validate_filename)
     _validate_regex = validator('regex', 'not_regex', allow_reuse=True)(validate_regex)
-    _validate_workdir = validator('workdir', allow_reuse=True)(validate_filename)
     _validate_tags = validator('tags', each_item=True, allow_reuse=True)(validate_catalog_tag)
+
+    @root_validator(skip_on_failure=True)
+    def mutex_validations(cls, values: Dict[str, Any]) -> Dict[str, Any]:
+        if bool(values.get('archive')) == bool(values.get('workdir')):
+            raise ValueError('Either "archive" or "workdir" must to be provided')
+
+        if values.get('regex') is not None and values.get('not_regex') is not None:
+            raise ValueError('Argument "not_regex" not allowed with "regex"')
+
+        return values