+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+Sastre 1.20 [xx, 2022]
+================================
+
+#### Enhancements:
+- [#42] Option --archive added to backup and restore tasks. Allowing backups to be saved as a zip file, and restore to read from a zip file.
+
+
 Sastre 1.19 [August 16, 2022]
 ================================